integrating truffle tests into the overall test suite
