+++ .travis.yml
@@ -9,7 +9,7 @@ services:
 
 env:
   global:
-    - ELASTICSEARCH=http://127.0.0.1:9200 
+    - ELASTICSEARCH=http://127.0.0.1:9200
     - PGHOST=localhost
     - PGPORT=5444
     - PGUSER=postgres
@@ -28,6 +28,8 @@ install:
   - travis_wait ./node_modules/.bin/lerna bootstrap
 
 before_script:
+ - yarn run blockchain &
+ - sleep 5
  - wget -q --waitretry=1 --retry-connrefused -T 1 -O - http://127.0.0.1:9200
  - node packages/postgresql/node-tests/pg-server/wait-for-db.js
