+++ packages/ethereum/.eslintrc.js
@@ -1,4 +1,4 @@
 module.exports = {
   root: true,
-  extends: '@cardstack/eslint-config/browser',
+  extends: '@cardstack/eslint-config/browser'
 };
