--- packages/ethereum/node-tests/.eslintrc.js
@@ -1,3 +1,0 @@
-module.exports = {
-  "extends": require.resolve("@cardstack/eslint-config/test")
-};
