+++ packages/ethereum/package.json
@@ -20,7 +20,9 @@
   "scripts": {
     "build": "ember build",
     "start": "ember serve",
-    "test": "ember try:each"
+    "test": "ember try:each",
+    "ethereum-tests": "../../node_modules/.bin/truffle test",
+    "private-blockchain": "../../node_modules/.bin/truffle develop"
   },
   "dependencies": {
     "@cardstack/logger": "^0.1.0",
