+++ packages/ethereum/truffle-tests/.eslintrc.js
@@ -1,9 +1,11 @@
 module.exports = {
   "extends": require.resolve("@cardstack/eslint-config/test"),
+  env: {
+    'node': true
+  },
   "globals": {
     "web3": true,
     "contract": true,
     "artifacts": true,
-    "assert": true
-  },
+  }
 };
