+++ packages/ethereum/truffle.js
@@ -1,5 +1,5 @@
 /*eslint-env node */
 
 module.exports = {
+  test_directory: "truffle-tests",
 };
-
